+++ magpie/api/schemas.py
@@ -23,7 +23,7 @@ from pyramid.security import NO_PERMISSION_REQUIRED
 
 from magpie import __meta__
 from magpie.constants import get_constant
-from magpie.permissions import Permission
+from magpie.permissions import Access, Permission, PermissionType, Scope
 from magpie.security import get_provider_names
 from magpie.utils import (
     CONTENT_TYPE_HTML,
@@ -584,7 +584,44 @@ class UserNamesListSchema(colander.SequenceSchema):
     user_name = UserNameParameter
 
 
-class PermissionListSchema(colander.SequenceSchema):
+class Permission_Check_BadRequestResponseSchema(BaseResponseSchemaAPI):
+    description = "Invalid permission format. Must be a permission string name or detailed JSON object."
+    body = ErrorResponseBodySchema(code=HTTPBadRequest.code, description=description)
+
+
+class PermissionObjectSchema(colander.MappingSchema):
+    name = colander.SchemaNode(
+        colander.String(),
+        description="Permission name applicable to the service/resource.",
+        example=Permission.READ.value
+    )
+    access = colander.SchemaNode(
+        colander.String(),
+        description="Permission access rule to the service/resource.",
+        default=Access.ALLOW.value,
+        example=Access.ALLOW.value,
+    )
+    scope = colander.SchemaNode(
+        colander.String(),
+        description="Permission scope over service/resource tree hierarchy.",
+        default=Scope.RECURSIVE.value,
+        example=Scope.RECURSIVE.value,
+    )
+
+
+class PermissionObjectTypeSchema(PermissionObjectSchema):
+    type = colander.SchemaNode(
+        colander.String(),
+        description="Permission type being displayed.",
+        example=PermissionType.ALLOWED.value
+    )
+
+
+class PermissionObjectListSchema(colander.SequenceSchema):
+    permission = PermissionObjectTypeSchema()
+
+
+class PermissionNameListSchema(colander.SequenceSchema):
     permission_name = colander.SchemaNode(
         colander.String(),
         description="Permissions applicable to the service/resource",
@@ -654,9 +691,12 @@ class ServiceBodySchema(colander.MappingSchema):
         colander.Integer(),
         description="Resource identification number",
     )
-    permission_names = PermissionListSchema(
+    permission_names = PermissionNameListSchema(
         description="List of service permissions applicable or effective for a given user/group according to context.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed service permissions applicable or effective for user/group according to context."
     )
     service_name = colander.SchemaNode(
         colander.String(),
@@ -717,9 +757,13 @@ class ResourceBodySchema(colander.MappingSchema):
         default=colander.null,  # if no parent
         missing=colander.drop   # if not returned (basic_info = True)
     )
-    permission_names = PermissionListSchema(example=[Permission.READ.value, Permission.WRITE.value],
-                                            description="List of resource permissions applicable or effective "
-                                                        "for a given user/group according to context.")
+    permission_names = PermissionNameListSchema(
+        description="List of resource permissions applicable or effective for a given user/group according to context.",
+        example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed resource permissions applicable or effective for user/group according to context."
+    )
     permission_names.default = colander.null  # if no parent
     permission_names.missing = colander.drop  # if not returned (basic_info = True)
 
@@ -934,9 +978,12 @@ class Resources_POST_ConflictResponseSchema(BaseResponseSchemaAPI):
 
 
 class ResourcePermissions_GET_ResponseBodySchema(BaseResponseBodySchema):
-    permission_names = PermissionListSchema(
+    permission_names = PermissionNameListSchema(
         description="List of permissions applicable for the referenced resource.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed permissions applicable for the referenced resource."
     )
 
 
@@ -1213,7 +1260,13 @@ class Service_DELETE_ForbiddenResponseSchema(BaseResponseSchemaAPI):
 
 
 class ServicePermissions_ResponseBodySchema(BaseResponseBodySchema):
-    permission_names = PermissionListSchema(example=[Permission.READ.value, Permission.WRITE.value])
+    permission_names = PermissionNameListSchema(
+        description="List of permissions applicable to the service.",
+        example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed permissions applicable to the service."
+    )
 
 
 class ServicePermissions_GET_OkResponseSchema(BaseResponseSchemaAPI):
@@ -1272,9 +1325,12 @@ class ServiceTypeResourceInfo(colander.MappingSchema):
         colander.Boolean(),
         description="Indicates if the resource type allows child resources."
     )
-    permission_names = PermissionListSchema(
+    permission_names = PermissionNameListSchema(
         description="Permissions applicable to the specific resource type.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed permissions applicable to the specific resource type."
     )
 
 
@@ -1670,9 +1726,12 @@ class UserResourcePermissions_GET_RequestSchema(BaseRequestSchemaAPI):
 
 
 class UserResourcePermissions_GET_ResponseBodySchema(BaseResponseBodySchema):
-    permission_names = PermissionListSchema(
-        description="List of resource permissions effective for the referenced user.",
+    permission_names = PermissionNameListSchema(
+        description="List of resource permissions applied or effective for the referenced user.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed resource permissions applied or effective to the referenced user."
     )
 
 
@@ -1718,7 +1777,13 @@ class UserResourcePermissions_GET_NotFoundResponseSchema(BaseResponseSchemaAPI):
 class UserResourcePermissions_POST_RequestBodySchema(colander.MappingSchema):
     permission_name = colander.SchemaNode(
         colander.String(),
-        description="permission_name of the created user-resource-permission reference.")
+        description="Permission name (implicit or explicit) to create on the resource for the user.",
+        missing=colander.drop
+    )
+    permission = PermissionObjectSchema(
+        description="Detailed permission definition to create on the resource for the user.",
+        missing=colander.drop
+    )
 
 
 class UserResourcePermissions_POST_RequestSchema(BaseRequestSchemaAPI):
@@ -1730,13 +1795,21 @@ class UserResourcePermissions_POST_RequestSchema(BaseRequestSchemaAPI):
 class UserResourcePermissions_POST_ResponseBodySchema(BaseResponseBodySchema):
     resource_id = colander.SchemaNode(
         colander.Integer(),
-        description="resource_id of the created user-resource-permission reference.")
+        description="Resource for which to create the user permission."
+    )
     user_id = colander.SchemaNode(
         colander.Integer(),
-        description="user_id of the created user-resource-permission reference.")
+        description="User for which to create the permission on the resource."
+    )
     permission_name = colander.SchemaNode(
         colander.String(),
-        description="permission_name of the created user-resource-permission reference.")
+        description="Permission name (implicit or explicit) to create on the resource for the user.",
+        missing=colander.drop
+    )
+    permission = PermissionObjectSchema(
+        description="Detailed permission definition to create on the resource for the user.",
+        missing=colander.drop
+    )
 
 
 class UserResourcePermissions_POST_CreatedResponseSchema(BaseResponseSchemaAPI):
@@ -1749,48 +1822,64 @@ class UserResourcePermissions_POST_ParamResponseBodySchema(colander.MappingSchem
     value = colander.SchemaNode(colander.String(), description="Specified parameter value.")
 
 
-class UserResourcePermissions_POST_BadResponseBodySchema(BaseResponseBodySchema):
+class UserResourcePermissions_POST_ErrorResponseBodySchema(ErrorResponseBodySchema):
     user_name = colander.SchemaNode(colander.String(), description="Specified user name.")
     resource_id = colander.SchemaNode(colander.String(), description="Specified resource id.")
     permission_name = colander.SchemaNode(colander.String(), description="Specified permission name.")
+    permission = PermissionObjectTypeSchema(description="Specific permission definition.")
     param = UserResourcePermissions_POST_ParamResponseBodySchema(missing=colander.drop)
 
 
 class UserResourcePermissions_POST_BadRequestResponseSchema(BaseResponseSchemaAPI):
     description = "Permission not allowed for specified 'resource_type'."
-    body = UserResourcePermissions_POST_BadResponseBodySchema(code=HTTPBadRequest.code, description=description)
+    body = UserResourcePermissions_POST_ErrorResponseBodySchema(code=HTTPBadRequest.code, description=description)
 
 
 class UserResourcePermissions_POST_ForbiddenResponseSchema(BaseResponseSchemaAPI):
     description = "Creation of permission on resource for user refused by db."
-    body = UserResourcePermissions_POST_BadResponseBodySchema(code=HTTPForbidden.code, description=description)
+    body = UserResourcePermissions_POST_ErrorResponseBodySchema(code=HTTPForbidden.code, description=description)
 
 
 class UserResourcePermissions_POST_ConflictResponseSchema(BaseResponseSchemaAPI):
     description = "Permission already exist on resource for user."
-    body = UserResourcePermissions_POST_ResponseBodySchema(code=HTTPConflict.code, description=description)
+    body = UserResourcePermissions_POST_ErrorResponseBodySchema(code=HTTPConflict.code, description=description)
 
 
 # using same definitions
-UserResourcePermissions_DELETE_BadResponseBodySchema = UserResourcePermissions_POST_ResponseBodySchema
-UserResourcePermissions_DELETE_BadRequestResponseSchema = UserResourcePermissions_POST_BadRequestResponseSchema
+UserResourcePermissionName_DELETE_ErrorResponseBodySchema = UserResourcePermissions_POST_ErrorResponseBodySchema
+UserResourcePermissionName_DELETE_BadRequestResponseSchema = UserResourcePermissions_POST_BadRequestResponseSchema
 
 
-class UserResourcePermission_DELETE_RequestSchema(BaseRequestSchemaAPI):
+class UserResourcePermissionName_DELETE_RequestSchema(BaseRequestSchemaAPI):
     body = colander.MappingSchema(default={})
     user_name = UserNameParameter
     resource_id = ResourceIdParameter
     permission_name = PermissionNameParameter
 
 
-class UserResourcePermissions_DELETE_OkResponseSchema(BaseResponseSchemaAPI):
+class UserResourcePermissionName_DELETE_OkResponseSchema(BaseResponseSchemaAPI):
     description = "Delete user resource permission successful."
     body = BaseResponseBodySchema(code=HTTPOk.code, description=description)
 
 
-class UserResourcePermissions_DELETE_NotFoundResponseSchema(BaseResponseSchemaAPI):
+class UserResourcePermissionName_DELETE_NotFoundResponseSchema(BaseResponseSchemaAPI):
     description = "Could not find user resource permission to delete from db."
-    body = UserResourcePermissions_DELETE_BadResponseBodySchema(code=HTTPOk.code, description=description)
+    body = UserResourcePermissionName_DELETE_ErrorResponseBodySchema(code=HTTPNotFound.code, description=description)
+
+
+class UserResourcePermissions_DELETE_RequestBodySchema(colander.MappingSchema):
+    permission = PermissionObjectSchema(description="Permission to be deleted.")
+
+
+class UserResourcePermissions_DELETE_RequestSchema(BaseRequestSchemaAPI):
+    body = UserResourcePermissions_DELETE_RequestBodySchema()
+    user_name = UserNameParameter
+    resource_id = ResourceIdParameter
+
+
+UserResourcePermissions_DELETE_OkResponseSchema = UserResourcePermissionName_DELETE_OkResponseSchema
+UserResourcePermissions_DELETE_BadRequestResponseSchema = UserResourcePermissionName_DELETE_BadRequestResponseSchema
+UserResourcePermissions_DELETE_NotFoundResponseSchema = UserResourcePermissionName_DELETE_NotFoundResponseSchema
 
 
 class UserServiceResources_GET_ResponseBodySchema(BaseResponseBodySchema):
@@ -1813,7 +1902,15 @@ class UserServiceResources_GET_RequestSchema(BaseRequestSchemaAPI):
 
 
 class UserServicePermissions_POST_RequestBodySchema(colander.MappingSchema):
-    permission_name = colander.SchemaNode(colander.String(), description="Name of the permission to create.")
+    permission_name = colander.SchemaNode(
+        colander.String(),
+        description="Permission name (implicit or explicit) to create on the service for the user.",
+        missing=colander.drop
+    )
+    permission = PermissionObjectSchema(
+        description="Detailed permission definition to create on the service for the user.",
+        missing=colander.drop
+    )
 
 
 class UserServicePermissions_POST_RequestSchema(BaseRequestSchemaAPI):
@@ -1822,7 +1919,17 @@ class UserServicePermissions_POST_RequestSchema(BaseRequestSchemaAPI):
     service_name = ServiceNameParameter
 
 
-class UserServicePermission_DELETE_RequestSchema(BaseRequestSchemaAPI):
+class UserServicePermissions_DELETE_RequestBodySchema(colander.MappingSchema):
+    permission = PermissionObjectSchema(description="Permission to be deleted.")
+
+
+class UserServicePermissions_DELETE_RequestSchema(BaseRequestSchemaAPI):
+    body = UserResourcePermissions_DELETE_RequestBodySchema()
+    user_name = UserNameParameter
+    resource_id = ResourceIdParameter
+
+
+class UserServicePermissionName_DELETE_RequestSchema(BaseRequestSchemaAPI):
     body = colander.MappingSchema(default={})
     user_name = UserNameParameter
     service_name = ServiceNameParameter
@@ -1860,9 +1967,12 @@ class UserServicePermissions_GET_RequestSchema(BaseRequestSchemaAPI):
 
 
 class UserServicePermissions_GET_ResponseBodySchema(BaseResponseBodySchema):
-    permission_names = PermissionListSchema(
-        description="List of service permissions effective for the referenced user.",
+    permission_names = PermissionNameListSchema(
+        description="List of service permissions applied or effective for the referenced user.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed service permissions applied or effective to the referenced user."
     )
 
 
@@ -2064,9 +2174,12 @@ class GroupServices_InternalServerErrorResponseSchema(BaseResponseSchemaAPI):
 
 
 class GroupServicePermissions_GET_ResponseBodySchema(BaseResponseBodySchema):
-    permission_names = PermissionListSchema(
-        description="List of service permissions effective for the referenced group.",
+    permission_names = PermissionNameListSchema(
+        description="List of service permissions applied to the referenced group.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed service permissions applied to the referenced group."
     )
 
 
@@ -2087,7 +2200,15 @@ class GroupServicePermissions_GET_InternalServerErrorResponseSchema(BaseResponse
 
 
 class GroupServicePermissions_POST_RequestBodySchema(colander.MappingSchema):
-    permission_name = colander.SchemaNode(colander.String(), description="Name of the permission to create.")
+    permission_name = colander.SchemaNode(
+        colander.String(),
+        description="Permission name (implicit or explicit) to create on the service for the group.",
+        missing=colander.drop
+    )
+    permission = PermissionObjectSchema(
+        description="Detailed permission definition to create on the service for the group.",
+        missing=colander.drop
+    )
 
 
 class GroupServicePermissions_POST_RequestSchema(BaseRequestSchemaAPI):
@@ -2104,6 +2225,7 @@ class GroupResourcePermissions_POST_RequestSchema(BaseRequestSchemaAPI):
 
 class GroupResourcePermissions_POST_ResponseBodySchema(BaseResponseBodySchema):
     permission_name = colander.SchemaNode(colander.String(), description="Name of the permission requested.")
+    permission = PermissionObjectTypeSchema(description="Detailed permission definition.")
     resource = ResourceBodySchema()
     group = GroupInfoBodySchema()
 
@@ -2133,11 +2255,24 @@ class GroupResourcePermissions_POST_ConflictResponseSchema(BaseResponseSchemaAPI
     body = GroupResourcePermissions_POST_ResponseBodySchema(code=HTTPConflict.code, description=description)
 
 
-class GroupResourcePermission_DELETE_RequestSchema(BaseRequestSchemaAPI):
+class GroupResourcePermissionName_DELETE_RequestSchema(BaseRequestSchemaAPI):
     body = colander.MappingSchema(default={})
     group_name = GroupNameParameter
     resource_id = ResourceIdParameter
     permission_name = PermissionNameParameter
+
+
+class GroupResourcePermissions_DELETE_RequestBodySchema(colander.MappingSchema):
+    permission = PermissionObjectSchema(
+        description="Detailed permission definition to delete from the resource for the group.",
+        missing=colander.drop
+    )
+
+
+class GroupResourcePermissions_DELETE_RequestSchema(BaseRequestSchemaAPI):
+    body = GroupResourcePermissions_DELETE_RequestBodySchema()
+    group_name = GroupNameParameter
+    resource_id = ResourceIdParameter
 
 
 class GroupResourcesPermissions_InternalServerErrorResponseBodySchema(InternalServerErrorResponseBodySchema):
@@ -2183,9 +2318,12 @@ class GroupResources_GET_InternalServerErrorResponseSchema(BaseResponseSchemaAPI
 
 
 class GroupResourcePermissions_GET_ResponseBodySchema(BaseResponseBodySchema):
-    permissions_names = PermissionListSchema(
-        description="List of resource permissions effective for the referenced group.",
+    permissions_names = PermissionNameListSchema(
+        description="List of resource permissions applied to the referenced group.",
         example=[Permission.READ.value, Permission.WRITE.value]
+    )
+    permissions = PermissionObjectListSchema(
+        description="List of detailed resource permissions applied to the referenced group."
     )
 
 
@@ -2203,11 +2341,14 @@ class GroupServiceResources_GET_OkResponseSchema(BaseResponseSchemaAPI):
     body = GroupServiceResources_GET_ResponseBodySchema(code=HTTPOk.code, description=description)
 
 
+GroupServicePermissions_DELETE_RequestSchema = GroupResourcePermissions_DELETE_RequestSchema
+
+
 class GroupServicePermission_DELETE_RequestBodySchema(colander.MappingSchema):
     permission_name = colander.SchemaNode(colander.String(), description="Name of the permission to delete.")
 
 
-class GroupServicePermission_DELETE_RequestSchema(BaseRequestSchemaAPI):
+class GroupServicePermissionName_DELETE_RequestSchema(BaseRequestSchemaAPI):
     body = GroupServicePermission_DELETE_RequestBodySchema()
     group_name = GroupNameParameter
     service_name = ServiceNameParameter
@@ -2215,9 +2356,17 @@ class GroupServicePermission_DELETE_RequestSchema(BaseRequestSchemaAPI):
 
 
 class GroupServicePermission_DELETE_ResponseBodySchema(BaseResponseBodySchema):
-    permission_name = colander.SchemaNode(colander.String(), description="Name of the permission requested.")
-    resource = ResourceBodySchema()
     group = GroupInfoBodySchema()
+    resource = ResourceBodySchema()
+    permission_name = colander.SchemaNode(
+        colander.String(),
+        description="Permission name (implicit or explicit) to be created.",
+        missing=colander.drop
+    )
+    permission = PermissionObjectSchema(
+        description="Detailed permission to be created",
+        missing=colander.drop
+    )
 
 
 class GroupServicePermission_DELETE_OkResponseSchema(BaseResponseSchemaAPI):
@@ -2777,11 +2926,20 @@ UserResourcePermissions_POST_responses = {
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
-UserResourcePermission_DELETE_responses = {
+UserResourcePermissions_DELETE_responses = {
     "200": UserResourcePermissions_DELETE_OkResponseSchema(),
     "400": UserResourcePermissions_DELETE_BadRequestResponseSchema(),
     "401": UnauthorizedResponseSchema(),
     "404": UserResourcePermissions_DELETE_NotFoundResponseSchema(),
+    "406": NotAcceptableResponseSchema(),
+    "422": UnprocessableEntityResponseSchema(),
+    "500": InternalServerErrorResponseSchema(),
+}
+UserResourcePermissionName_DELETE_responses = {
+    "200": UserResourcePermissionName_DELETE_OkResponseSchema(),
+    "400": UserResourcePermissionName_DELETE_BadRequestResponseSchema(),
+    "401": UnauthorizedResponseSchema(),
+    "404": UserResourcePermissionName_DELETE_NotFoundResponseSchema(),
     "406": NotAcceptableResponseSchema(),
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
@@ -2814,7 +2972,8 @@ UserServiceResources_GET_responses = {
     "500": InternalServerErrorResponseSchema(),
 }
 UserServicePermissions_POST_responses = UserResourcePermissions_POST_responses
-UserServicePermission_DELETE_responses = UserResourcePermission_DELETE_responses
+UserServicePermissions_DELETE_responses = UserResourcePermissions_DELETE_responses
+UserServicePermissionName_DELETE_responses = UserResourcePermissionName_DELETE_responses
 LoggedUser_GET_responses = {
     "200": User_GET_OkResponseSchema(),
     "403": User_CheckAnonymous_ForbiddenResponseSchema(),
@@ -2894,11 +3053,20 @@ LoggedUserResourcePermissions_POST_responses = {
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
-LoggedUserResourcePermission_DELETE_responses = {
+LoggedUserResourcePermissions_DELETE_responses = {
     "200": UserResourcePermissions_DELETE_OkResponseSchema(),
     "400": UserResourcePermissions_DELETE_BadRequestResponseSchema(),
     "401": UnauthorizedResponseSchema(),
     "404": UserResourcePermissions_DELETE_NotFoundResponseSchema(),
+    "406": NotAcceptableResponseSchema(),
+    "422": UnprocessableEntityResponseSchema(),
+    "500": InternalServerErrorResponseSchema(),
+}
+LoggedUserResourcePermissionName_DELETE_responses = {
+    "200": UserResourcePermissionName_DELETE_OkResponseSchema(),
+    "400": UserResourcePermissionName_DELETE_BadRequestResponseSchema(),
+    "401": UnauthorizedResponseSchema(),
+    "404": UserResourcePermissionName_DELETE_NotFoundResponseSchema(),
     "406": NotAcceptableResponseSchema(),
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
@@ -2928,7 +3096,8 @@ LoggedUserServiceResources_GET_responses = {
     "500": InternalServerErrorResponseSchema(),
 }
 LoggedUserServicePermissions_POST_responses = LoggedUserResourcePermissions_POST_responses
-LoggedUserServicePermission_DELETE_responses = LoggedUserResourcePermission_DELETE_responses
+LoggedUserServicePermissions_DELETE_responses = LoggedUserResourcePermissions_DELETE_responses
+LoggedUserServicePermissionName_DELETE_responses = LoggedUserResourcePermissionName_DELETE_responses
 Groups_GET_responses = {
     "200": Groups_GET_OkResponseSchema(),
     "401": UnauthorizedResponseSchema(),
@@ -3021,7 +3190,7 @@ GroupResourcePermissions_POST_responses = {
     "500": InternalServerErrorResponseSchema(),
 }
 GroupServicePermissions_POST_responses = GroupResourcePermissions_POST_responses
-GroupServicePermission_DELETE_responses = {
+GroupServicePermissionName_DELETE_responses = {
     "200": GroupServicePermission_DELETE_OkResponseSchema(),
     "401": UnauthorizedResponseSchema(),
     "403": GroupServicePermission_DELETE_ForbiddenResponseSchema(),
@@ -3030,6 +3199,9 @@ GroupServicePermission_DELETE_responses = {
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
+GroupServicePermissions_DELETE_responses = GroupServicePermissionName_DELETE_responses
+GroupResourcePermissions_DELETE_responses = GroupServicePermissions_DELETE_responses
+GroupResourcePermissionName_DELETE_responses = GroupServicePermissionName_DELETE_responses
 GroupResources_GET_responses = {
     "200": GroupResources_GET_OkResponseSchema(),
     "401": UnauthorizedResponseSchema(),
@@ -3048,7 +3220,6 @@ GroupResourcePermissions_GET_responses = {
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
-GroupResourcePermission_DELETE_responses = GroupServicePermission_DELETE_responses
 RegisterGroups_GET_responses = {
     "200": RegisterGroups_GET_OkResponseSchema(),
     "401": UnauthorizedResponseSchema(),